+++ srcs/vec_math/vec2_b.c
@@ -6,7 +6,7 @@
 /*   By: jnivala <jnivala@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/05 19:16:10 by tmaarela          #+#    #+#             */
-/*   Updated: 2021/02/02 16:37:09 by jnivala          ###   ########.fr       */
+/*   Updated: 2021/02/19 14:38:26 by jnivala          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 t_xy			vec2_rot(t_xy v, float angle)
 {
 	return ((t_xy){
-		v.x * cos(angle) - v.y * sin(angle),
-		v.x * sin(angle) + v.y * cos(angle)
+		v.x * cosf(angle) - v.y * sinf(angle),
+		v.x * sinf(angle) + v.y * cosf(angle)
 	});
 }
 
@@ -24,7 +24,7 @@ t_xy			vec2_norm(t_xy v)
 {
 	float mag;
 
-	mag = sqrt((v.x * v.x) + (v.y * v.y));
+	mag = sqrtf((v.x * v.x) + (v.y * v.y));
 	return ((t_xy){
 		v.x / mag,
 		v.y / mag
@@ -46,12 +46,12 @@ float			vec2_dot(t_xy a, t_xy b)
 
 float			vec2_mag(t_xy a)
 {
-	return sqrt(a.x * a.x + a.y * a.y);
+	return sqrtf(a.x * a.x + a.y * a.y);
 }
 
 float			vec2_angle(t_xy a, t_xy b)
 {
-	return acos(vec2_dot(a, b) / (vec2_mag(a) * vec2_mag(b)));
+	return acosf(vec2_dot(a, b) / (vec2_mag(a) * vec2_mag(b)));
 }
 
 float			ft_map(float in, t_range from, t_range to)
